+++ packages/runtime/src/index.ts
@@ -2,6 +2,7 @@ export * from "./RuntimeEvent";
 export * from "./RuntimeHostConfiguration";
 export * from "./RuntimeHost";
 export * from "./RuntimeConfigurationValidator";
+export * from "./RuntimeDiagnosticEvent";
 export * from "./RuntimeDiagnosticIssueCode";
 export * from "./RuntimeModule";
 export * from "./RuntimeHealthReport";